requirements.txt
psd-tools>=1.9.0
Pillow>=10.0.0
PySide6>=6.5.0
pyinstaller>=6.0.0
